refactor: split DocxToolbar into focused modules
Break the 1300-line DocxToolbar.tsx into a DocxToolbar/ directory with
single-purpose files: shared styles, the portaled Menu, TableGridPicker,
one component per tool group (each owning its local UI state, e.g. the
zoom input and font search query), the right-side ToolbarActions region,
and two hooks — useEditorFormatState (selection/zoom sync with the live
editor) and useToolbarOverflow (responsive tail-first collapse). The
main index composes them; public props and behavior are unchanged.

One robustness tweak: the action-region measurement now uses a
callback-ref instead of a plain ref with a prop dependency list, so a
remounting region is re-observed automatically.
